+++ extensions/cornerstone/src/services/ToolGroupService/ToolGroupService.ts
@@ -1,6 +1,7 @@
 import { ToolGroupManager, Enums, Types } from '@cornerstonejs/tools';
 
 import { Types as OhifTypes, pubSubServiceInterface } from '@ohif/core';
+import getActiveViewportEnabledElement from '../../utils/getActiveViewportEnabledElement';
 
 const EVENTS = {
   VIEWPORT_ADDED: 'event::cornerstone::toolgroupservice:viewportadded',
@@ -39,20 +40,56 @@ export default class ToolGroupService {
   EVENTS: { [key: string]: string };
 
   constructor(serviceManager) {
-    const { cornerstoneViewportService } = serviceManager.services;
+    const {
+      cornerstoneViewportService,
+      viewportGridService,
+    } = serviceManager.services;
     this.cornerstoneViewportService = cornerstoneViewportService;
+    this.viewportGridService = viewportGridService;
     this.listeners = {};
     this.EVENTS = EVENTS;
     Object.assign(this, pubSubServiceInterface);
   }
 
   /**
-   * Returns the cornerstone ToolGroup for a given toolGroup UID
-   * @param {string} toolGroupId - The toolGroup uid
-   * @returns {IToolGroup} - The toolGroup
+   * Retrieves a tool group from the ToolGroupManager by tool group ID.
+   * If no tool group ID is provided, it retrieves the tool group of the active viewport.
+   * @param toolGroupId - Optional ID of the tool group to retrieve.
+   * @returns The tool group or undefined if it is not found.
    */
-  public getToolGroup(toolGroupId: string): Types.IToolGroup | void {
-    const toolGroup = ToolGroupManager.getToolGroup(toolGroupId);
+  public getToolGroup(toolGroupId?: string): Types.IToolGroup | void {
+    let toolGroupIdToUse = toolGroupId;
+
+    if (!toolGroupIdToUse) {
+      // Use the active viewport's tool group if no tool group id is provided
+      const enabledElement = getActiveViewportEnabledElement(
+        this.viewportGridService
+      );
+
+      if (!enabledElement) {
+        return;
+      }
+
+      const { renderingEngineId, viewportId } = enabledElement;
+      const toolGroup = ToolGroupManager.getToolGroupForViewport(
+        viewportId,
+        renderingEngineId
+      );
+
+      if (!toolGroup) {
+        console.warn(
+          'No tool group found for viewportId:',
+          viewportId,
+          'and renderingEngineId:',
+          renderingEngineId
+        );
+        return;
+      }
+
+      toolGroupIdToUse = toolGroup.id;
+    }
+
+    const toolGroup = ToolGroupManager.getToolGroup(toolGroupIdToUse);
     return toolGroup;
   }
